Matter can cross the edge of the slide, and a channel finds its level
`Scenario::flux`: a list of `Source` and `Drain` rectangles applied once
per fluid step. What it makes describable is the three habitats a closed
slide cannot be — a deep-sea vent is a slide in the dark with inorganic
matter welling up, marine snow is one with organic matter falling
through, and a channel is one with both ends open. `the_drift.ron` is now
the third.

**A source without a drain is a slow count up to the quantity cap**, and
a population under a ceiling set by arithmetic is not a population under
a carrying capacity — the same degenerate answer §17.7 finds for the
energy economy, reached from the other side. With both ends the standing
stock has a fixed point *in the stock itself*, because a drain is a `Q10`
fraction per step rather than an amount: what leaves scales with what is
there, which is exactly the `S`-dependent term §17.7 shows energy lacks.

Measured on the drift over six thousand ticks: detritus starts at 909M,
influx steady at 336M per thousand, efflux falls 445M → 330M as the
channel drains down, and the stock settles at **733M** and holds.
Population peaks at 89 and settles at 58. In a bare 32-square box the
fixed point is exact — 165,155,680 for four hundred ticks running, not
drifting — because integer truncation gives it a genuine dead band.

Both directions go through the ledger in both currencies, because both
invariants are identities and not approximations. Matter: what
`add_chem` reports as *actually* moved is what is recorded, so a source
aimed at a barrier or at a square already at the cap records what it
managed. Rectangles clamp to the slide rather than wrap — `index`
reduces modulo the grid, so an inlet running off the right-hand edge
would reappear at the outflow. Energy: matter that is a metabolic
substrate carries its latent energy with it, and `latent_in` weighs a
parcel by the same rule `recompute_stored` weighs the world, so the two
cannot disagree. Detritus and carrion weigh nothing — they have to be
taken apart before they are worth anything.

`TrophicSource::Influx` rather than folding it into `Light`, or a vent
slide in the dark would report a photosynthetic economy.

**Energy leaving with matter is not dissipation.** `energy_out` carries
both; `energy_exported` is the part that left latent in matter, and
`Sample::dissipation` now subtracts it, so dissipation goes on meaning
heat. A budget panel showing food washing off the downstream end as the
world getting warmer would describe a different world. `influx` and
`efflux` join it as rates, differenced here rather than in a panel for
the reason `absorbed` is.

Neither end touches cells. A cell that swims into an outflow is not
deleted: "the current washed it off the slide" and "it was eaten by the
edge of the world" are different claims and only one is one this engine
should make.

Snapshot format 8 → 9 for the four new counters (hard rule 7), and
`Ledger::restore` takes a struct now rather than eleven positional
arguments — four of them are `[i64; CHEM_COUNT]` and two of those are
adjacent, which is a transposition that would compile, restore silently
and surface as a conservation failure thousands of ticks into a resumed
run. Same reasoning as `FluidRates`.

Ten tests, including that a chemical nothing eats brings no energy with
it — the one that catches weighing every chemical instead of the right
ones — and that a resumed run keeps stepping to the same state hash.

+++ crates/mm-core/tests/predator_probe.rs
@@ -429,7 +429,7 @@ fn a_predator_standing_on_its_food() {
                             );
                             world
                                 .ledger_mut()
-                                .record_injected(mm_core::ecology::CARRION, added);
+                                .record_injected(mm_core::ecology::CARRION, i64::from(added));
                             let after = world.substrate().chem_at(mm_core::ecology::CARRION, x, y);
                             let _ = (before, added, after);
                         }
@@ -899,7 +899,10 @@ fn what_bar_the_ancestor_can_clear() {
     let path = Path::new(env!("CARGO_MANIFEST_DIR")).join("../../genomes/ancestor.mm");
     let source = std::fs::read_to_string(&path).expect("genome source");
     eprintln!("\nancestor divide threshold, one founder, soup.ron:");
-    eprintln!("{:>10} {:>10} {:>10} {:>9}", "threshold", "pop @2400", "pop @6000", "energy");
+    eprintln!(
+        "{:>10} {:>10} {:>10} {:>9}",
+        "threshold", "pop @2400", "pop @6000", "energy"
+    );
     for threshold in [200u32, 140, 100, 70, 50, 30] {
         let src = source.replace(
             "        IMM     200\n        CMP",
